Fix the bug that would occur with a literal PUT body
The problem is the read callback function used for
reading the data. Somehow it hangs curl midflight
if there is too much data given. Also, it is really
meant for files - and it turns out you can just pass
the string pointer curl explicitly, as described
in this SO answer:

http://stackoverflow.com/a/7570281/153886

Also, switch to POSTFIELDSIZE_LARGE if it is
supported.

Closes #12