Allow setting download_byte_limit
It can be useful when working with downloads
from untrusted sources in combination with
calls like `get_file`. If we can limit the
amount of data received we can prevent servers
suddenly sending us tons and tons of data whereas
in fact we only expect a few KB. By default
the option is disabled.

If the download limit has been reached, fire the interrupt.
Store the interrupt reason in the state struct, since
we might abort the request due to different conditions.

To do so we use CURLOPT_XFERFUNCTION when available,
which is supposed to be the modern way to do the same,
with better guarantees that it _is_ going to be called.

Update interrupt handling, add Patron::Aborted
for when the callback tells CURL to stop.

Define the data write correctly for file downloads
for Windows compatibility (as per CURL doc).